requirements-test.txt
coverage>=5
flake8>=3.5
ruff>=0.4
pytest>=6.2.5
pytest-cov>=3.0.0
pytest-mock>=3.6.1
pytest-xdist>=1.34.0
pytest-rerunfailures>=10.2
pytest-timeout>=2.0.2
testfixtures>=6.18.5
mock>=4.0.3
responses>=0.20.0
PyYAML>=6.0
pytest-socket>=0.7
